Correcoes de funcionalidades da fase 3. Ainda existem outras que serao negociadas pois ha divergencia no que foi pedido.

+++ plastecno-service/src/br/com/plastecno/service/constante/FormaMaterial.java
@@ -25,8 +25,8 @@ public enum FormaMaterial {
 		return this.ipi;
 	}
 	
-	public double getIpiPercentual() {
-		return this.getIpi() * 100;
+	public int getIpiPercentual() {
+		return (int)(this.getIpi() * 100);
 	}
 	
 	public boolean contemLargura(){
